move SquareToolCard layouts to their own source file
